Added Replay System

+++ phase2/thegame/app/src/main/java/com/group0565/tsu/game/ArchiveInputEvent.java
@@ -1,11 +1,11 @@
 package com.group0565.tsu.game;
 
 public class ArchiveInputEvent {
-    public float position;
+    public double position;
     public long startTime;
     public long endTime;
 
-    public ArchiveInputEvent(float position, long startTime, long endTime) {
+    public ArchiveInputEvent(double position, long startTime, long endTime) {
         this.position = position;
         this.startTime = startTime;
         this.endTime = endTime;
